requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pyyaml==6.0.2
httpx==0.28.1
aiofiles==24.1.0
pytest==8.3.4
pytest-asyncio==0.25.3
